[Bronze V] Title: 이상한 기호, Time: 132 ms, Memory: 20692 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.kt
@@ -2,6 +2,6 @@ import java.io.BufferedReader
 import java.io.InputStreamReader
 
 fun main(args: Array<String>) = with (BufferedReader(InputStreamReader(System.`in`))) {
-    println( readLine().split(' ').map { it.toInt() }.let { (it[0] + it[1]) * (it[0] - it[1]) })
+    println( readLine().split(' ').map { it.toBigInteger() }.let { (it[0] + it[1]) * (it[0] - it[1]) })
 
 }
